backend/requirements.txt
# This is the master file; see deploy_requirements.txt for production.
# If necessary we could have a core set of requirements, and source
# that in for differences between prod/testing (eg, psycopg2 only in
# prod, and django-extensions only for testing)

Django==2.1.7
djangorestframework==3.9.2
django-extensions==2.1.6
pytz==2018.9
lxml==4.3.2
django-uuidfield==0.5.0
xmlunittest==0.5.0
django-bootstrap3==11.0.0
jsonfield==2.0.2
djangorestframework-jsonp==1.0.2
djangorestframework-xml==1.4.0
Pillow==5.4.1
django-imagekit==4.0.2
django-fsm==2.6.0
django-fsm-admin==1.2.4
requests==2.21.0
django-filter==2.1.0
django-treebeard==4.3
rdflib==4.2.2
mozilla-django-oidc==1.2.2

